refactor: 클래스명 수정, 리프레시토큰의 경우 id 초기화하지않도록 수정
